fix: move compose actions out of shared folder
