+++ config/compose.config.js
@@ -1,5 +1,35 @@
+const StyleDictionary = require('style-dictionary');
+const Color = require('tinycolor2');
+
 const { PREFIX, OUTPUT_PATH, OUTPUT_BASE_FILENAME } = process.env;
 const WHITELABEL = process.env.WHITELABEL !== 'false';
+
+
+StyleDictionary.registerTransform({
+  type: 'value',
+  name: 'color/composeColor',
+  matcher: (token) => token.original.type === 'color',
+  transformer: (token) => {
+    const hex8 = Color(token.value).toHex8();
+    return `Color(0x${hex8})`;
+  }
+});
+
+StyleDictionary.registerTransform({
+  type: 'value',
+  name: 'has-alpha',
+  transitive: true, 
+  transformer: function (token) {
+    if (token.value?.alpha != null) {
+      const hex = Color(token.value.color.replace('Color(0x', '').slice(0, -3))
+      const hex8withAlpha = hex.setAlpha(token.value.alpha)
+      token.value = `Color(0x${hex8withAlpha.toHex8()})`;
+      return token.value
+    } else {
+      return token.value;      
+    }
+  },
+});
 
 const composeObjectTransformGroup = [
     'has-alpha',
